+++ src/main/java/application/rest/UserController.java
@@ -12,7 +12,7 @@ import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequestMapping(value = "/users")
-public class UserController extends AbstractDatabaseController<User, Long, UserDTO, UserService> {
+public class UserController extends AbstractDatabaseController<User, Integer, UserDTO, UserService> {
 
     @Autowired
     private UserService userService;
@@ -23,7 +23,7 @@ public class UserController extends AbstractDatabaseController<User, Long, UserD
     }
 
     @RequestMapping(value = "/{id}", method = RequestMethod.GET)
-    public ResponseEntity<?> readUser(@PathVariable Long id) {
+    public ResponseEntity<?> readUser(@PathVariable Integer id) {
         return read(id);
     }
