+++ src/main/java/application/service/user/UserService.java
@@ -5,7 +5,7 @@ import application.rest.domain.UserDTO;
 import application.service.DatabaseServiceInterface;
 import application.service.response.ServiceResponse;
 
-public interface UserService extends DatabaseServiceInterface<User, Long, UserDTO> {
+public interface UserService extends DatabaseServiceInterface<User, Integer, UserDTO> {
 
     ServiceResponse<UserDTO> read(String username);
     ServiceResponse<UserDTO> create(UserDTO user);
